Change view.py into a package with multiple smaller files
