+++ gaphas/view/__init__.py
@@ -1,0 +1,5 @@
+"""This package contains everything to display a Canvas on a screen."""
+
+from gaphas.view.gtkview import DEFAULT_CURSOR, GtkView
+from gaphas.view.selection import Selection
+from gaphas.view.view import View
