+++ gaphas/view/gtkview.py
@@ -1,13 +1,12 @@
-"""This module contains everything to display a Canvas on a screen."""
 import cairo
 from gi.repository import Gdk, GLib, GObject, Gtk
 
 from gaphas.canvas import Context, instant_cairo_context
 from gaphas.decorators import AsyncIO
 from gaphas.geometry import Rectangle, distance_point_point_fast
-from gaphas.painter import BoundingBoxPainter, DefaultPainter, ItemPainter
-from gaphas.quadtree import Quadtree
 from gaphas.tool import DefaultTool
+from gaphas.view.selection import Selection
+from gaphas.view.view import View
 
 # Handy debug flag for drawing bounding boxes around the items.
 DEBUG_DRAW_BOUNDING_BOX = False
@@ -15,253 +14,6 @@ DEBUG_DRAW_QUADTREE = False
 
 # The default cursor (use in case of a cursor reset)
 DEFAULT_CURSOR = Gdk.CursorType.LEFT_PTR
-
-
-class View:
-    """View class for gaphas.Canvas objects."""
-
-    def __init__(self, canvas=None):
-        self._matrix = cairo.Matrix()
-        self._painter = DefaultPainter(self)
-        self._bounding_box_painter = BoundingBoxPainter(ItemPainter(self), self)
-
-        self._qtree = Quadtree()
-        self._bounds = Rectangle(0, 0, 0, 0)
-
-        self._canvas = None
-        if canvas:
-            self._set_canvas(canvas)
-
-    matrix = property(lambda s: s._matrix, doc="Canvas to view transformation matrix")
-
-    def _set_canvas(self, canvas):
-        """
-        Use view.canvas = my_canvas to set the canvas to be rendered
-        in the view.
-        """
-        if self._canvas:
-            self._qtree.clear()
-
-        self._canvas = canvas
-
-    canvas = property(lambda s: s._canvas, _set_canvas)
-
-    def queue_draw_item(self, *items):
-        """Placeholder for item redraw queueing."""
-        pass
-
-    def _set_painter(self, painter):
-        """Set the painter to use.
-
-        Painters should implement painter.Painter.
-        """
-        self._painter = painter
-        painter.set_view(self)
-
-    painter = property(lambda s: s._painter, _set_painter)
-
-    def _set_bounding_box_painter(self, painter):
-        """Set the painter to use for bounding box calculations."""
-        self._bounding_box_painter = painter
-        painter.set_view(self)
-
-    bounding_box_painter = property(
-        lambda s: s._bounding_box_painter, _set_bounding_box_painter
-    )
-
-    def get_port_at_point(self, vpos, distance=10, exclude=None):
-        """Find item with port closest to specified position.
-
-        List of items to be ignored can be specified with `exclude`
-        parameter.
-
-        Tuple is returned
-
-        - found item
-        - closest, connectable port
-        - closest point on found port (in view coordinates)
-
-        :Parameters:
-         vpos
-            Position specified in view coordinates.
-         distance
-            Max distance from point to a port (default 10)
-         exclude
-            Set of items to ignore.
-        """
-        v2i = self.get_matrix_v2i
-        vx, vy = vpos
-
-        max_dist = distance
-        port = None
-        glue_pos = None
-        item = None
-
-        rect = (vx - distance, vy - distance, distance * 2, distance * 2)
-        items = reversed(self.get_items_in_rectangle(rect))
-        for i in items:
-            if i in exclude:
-                continue
-            for p in i.ports():
-                if not p.connectable:
-                    continue
-
-                ix, iy = v2i(i).transform_point(vx, vy)
-                pg, d = p.glue((ix, iy))
-
-                if d >= max_dist:
-                    continue
-
-                max_dist = d
-                item = i
-                port = p
-
-                # transform coordinates from connectable item space to view
-                # space
-                i2v = self.get_matrix_i2v(i).transform_point
-                glue_pos = i2v(*pg)
-
-        return item, port, glue_pos
-
-    def get_items_in_rectangle(self, rect, intersect=True):
-        """Return the items in the rectangle 'rect'.
-
-        Items are automatically sorted in canvas' processing order.
-        """
-        assert self._canvas
-        if intersect:
-            items = self._qtree.find_intersect(rect)
-        else:
-            items = self._qtree.find_inside(rect)
-        return self._canvas.sort(items)
-
-    def zoom(self, factor):
-        """Zoom in/out by factor @factor."""
-        # TODO: should the scale factor be clipped?
-        self._matrix.scale(factor, factor)
-
-    def set_item_bounding_box(self, item, bounds):
-        """Update the bounding box of the item.
-
-        ``bounds`` is in view coordinates.
-
-        Coordinates are calculated back to item coordinates, so
-        matrix-only updates can occur.
-        """
-        v2i = self.get_matrix_v2i(item).transform_point
-        ix0, iy0 = v2i(bounds.x, bounds.y)
-        ix1, iy1 = v2i(bounds.x1, bounds.y1)
-        self._qtree.add(item=item, bounds=bounds, data=(ix0, iy0, ix1, iy1))
-
-    def get_item_bounding_box(self, item):
-        """Get the bounding box for the item, in view coordinates."""
-        return self._qtree.get_bounds(item)
-
-    bounding_box = property(lambda s: s._bounds)
-
-    def update_bounding_box(self, cr, items=None):
-        """Update the bounding boxes of the canvas items for this view, in
-        canvas coordinates."""
-        painter = self._bounding_box_painter
-        if items is None:
-            items = self.canvas.get_all_items()
-
-        # The painter calls set_item_bounding_box() for each rendered item.
-        painter.paint(Context(cairo=cr, items=items))
-
-        # Update the view's bounding box with the rest of the items
-        self._bounds = Rectangle(*self._qtree.soft_bounds)
-
-    def get_matrix_i2v(self, item):
-        """Get Item to View matrix for ``item``."""
-        matrix_i2c = self.canvas.get_matrix_i2c(item)
-        return matrix_i2c.multiply(self._matrix)
-
-    def get_matrix_v2i(self, item):
-        """Get View to Item matrix for ``item``."""
-        m = self.get_matrix_i2v(item)
-        m.invert()
-        return m
-
-
-class Selection(GObject.Object):
-    # Just defined a name to make GTK register this class.
-    __gtype_name__ = "GaphasSelection"
-
-    # Signals: emitted after the change takes effect.
-    __gsignals__ = {
-        "dropzone-changed": (
-            GObject.SignalFlags.RUN_LAST,
-            None,
-            (GObject.TYPE_PYOBJECT,),
-        ),
-        "hover-changed": (GObject.SignalFlags.RUN_LAST, None, (GObject.TYPE_PYOBJECT,)),
-        "focus-changed": (GObject.SignalFlags.RUN_LAST, None, (GObject.TYPE_PYOBJECT,)),
-        "selection-changed": (
-            GObject.SignalFlags.RUN_LAST,
-            None,
-            (GObject.TYPE_PYOBJECT,),
-        ),
-    }
-
-    def __init__(self, view):
-        super().__init__()
-        self._view = view
-        # Handling selections.
-        self._selected_items = set()
-        self._focused_item = None
-        self._hovered_item = None
-        self._dropzone_item = None
-
-    def clear(self):
-        self._selected_items.clear()
-        self._focused_item = None
-        self._hovered_item = None
-        self._dropzone_item = None
-
-    @property
-    def selected_items(self):
-        return self._selected_items
-
-    @property
-    def focused_item(self):
-        return self._focused_item
-
-    @property
-    def hovered_item(self):
-        return self._hovered_item
-
-    @property
-    def dropzone_item(self):
-        return self._dropzone_item
-
-    def add_selected_item(self, item):
-        if item not in self._selected_items:
-            self._selected_items.add(item)
-            self.emit("selection-changed", self._selected_items)
-
-    def remove_selected_item(self, item):
-        if item in self._selected_items:
-            self._selected_items.discard(item)
-            self.emit("selection-changed", self._selected_items)
-
-    def set_focused_item(self, item):
-        if item:
-            self.add_selected_item(item)
-
-        if item is not self._focused_item:
-            self._focused_item = item
-            self.emit("focus-changed", item)
-
-    def set_hovered_item(self, item):
-        if item is not self._hovered_item:
-            self._hovered_item = item
-            self.emit("hover-changed", item)
-
-    def set_dropzone_item(self, item):
-        if item is not self._dropzone_item:
-            self._dropzone_item = item
-            self.emit("dropzone-changed", item)
 
 
 class GtkView(Gtk.DrawingArea, Gtk.Scrollable, View):
